refactoring(2단계 - 로또(자동)) : 패키지 분류 수정
- service 레이어 삭제

- LottoNumber 패키지 이동
