+++ src/main/java/lotto/domain/LottoResult.java
@@ -1,4 +1,4 @@
-package lotto.service;
+package lotto.domain;
 
 import lotto.domain.Lotto;
 import lotto.domain.LottoTickets;
@@ -7,7 +7,7 @@ import lotto.domain.Rank;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class LottoResultService {
+public class LottoResult {
 
     public List<Rank> rankResult(LottoTickets lottoTickets) {
         return lottoTickets.getLottoList()
